Retrieving the protein sequence of human preproinsulin
Accessed NCBI at https://ncbi.nlm.nih.gov. selected human insulin is protein menu. searched human insulin. at the bottom of the search record, copied the insulin sequence, which starts with the word ORIGIN and ends with // . Pasted the insulin sequence into preproinsulin-seq.txt. Used regex to programmatically strip the file of ORIGIN, its numbers, the two slashes (//), spaces, and line breaks or return carriages. Confirmed file has 110 characters of lowercase letters, which represent the amino acids in the sequence of human preproinsulin with a character count. then saved certain characters in order to separate files text files. These text files are provided and so is the python file used to optimize the efficiency and speed to retrieve the information and partition it.